% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDTVARMx-methods-fit-dt-var-id-mx.R
\name{coef.dtvaridmx}
\alias{coef.dtvaridmx}
\title{Parameter Estimates}
\usage{
\method{coef}{dtvaridmx}(object, alpha = FALSE, psi = FALSE, theta = FALSE, ...)
}
\arguments{
\item{object}{Object of class \code{dtvaridmx}.}

\item{alpha}{Logical.
If \code{alpha = TRUE},
include estimates of the \code{alpha} vector, if available.
If \code{alpha = FALSE},
exclude estimates of the \code{alpha} vector.}

\item{psi}{Logical.
If \code{psi = TRUE},
include estimates of the \code{psi} matrix, if available.
If \code{psi = FALSE},
exclude estimates of the \code{psi} matrix.}

\item{theta}{Logical.
If \code{theta = TRUE},
include estimates of the \code{theta} matrix, if available.
If \code{theta = FALSE},
exclude estimates of the \code{theta} matrix.}

\item{...}{additional arguments.}
}
\value{
Returns a list of vectors of parameter estimates.
}
\description{
Parameter Estimates
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
