% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDTVARMx-methods-fit-dt-var-mx.R
\name{coef.fitdtvarmx}
\alias{coef.fitdtvarmx}
\title{Parameter Estimates}
\usage{
\method{coef}{fitdtvarmx}(object, psi = FALSE, theta = FALSE, ...)
}
\arguments{
\item{object}{Object of class \code{fitdtvarmx}.}

\item{psi}{Logical.
If \code{psi = TRUE},
include estimates of the \code{psi} matrix.
If \code{psi = FALSE},
exclude estimates of the \code{psi} matrix.}

\item{theta}{Logical.
If \code{theta = TRUE},
include estimates of the \code{theta} matrix if available.
If \code{theta = FALSE},
exclude estimates of the \code{theta} matrix.}

\item{...}{additional arguments.}
}
\value{
Returns a vector of parameter estimates.
}
\description{
Parameter Estimates
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
