% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDTVARMx-package.R
\docType{package}
\name{fitDTVARMx-package}
\alias{fitDTVARMx}
\alias{fitDTVARMx-package}
\title{fitDTVARMx: Fit The Discrete-Time Vector Autoregressive Model}
\description{
Fit the discrete-time vector autoregressive model using the 'OpenMx' package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jeksterslab/fitDTVARMx}
  \item \url{https://jeksterslab.github.io/fitDTVARMx/}
  \item Report bugs at \url{https://github.com/jeksterslab/fitDTVARMx/issues}
}

}
\author{
\strong{Maintainer}: Ivan Jacob Agaloos Pesigan \email{r.jeksterslab@gmail.com} (\href{https://orcid.org/0000-0003-4818-8420}{ORCID}) [copyright holder]

}
\keyword{internal}
