% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDTVARMx-methods-fit-dt-var-id-mx.R
\name{print.dtvaridmx}
\alias{print.dtvaridmx}
\title{Print Method for Object of Class \code{dtvaridmx}}
\usage{
\method{print}{dtvaridmx}(x, means = TRUE, ...)
}
\arguments{
\item{x}{an object of class \code{dtvaridmx}.}

\item{means}{Logical.
If \code{means = TRUE}, return means.
Otherwise, the function returns raw estimates.}

\item{...}{further arguments.}
}
\description{
Print Method for Object of Class \code{dtvaridmx}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
