% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDTVARMx-methods-fit-dt-var-mx.R
\name{print.dtvarmx}
\alias{print.dtvarmx}
\title{Print Method for Object of Class \code{dtvarmx}}
\usage{
\method{print}{dtvarmx}(x, ...)
}
\arguments{
\item{x}{an object of class \code{dtvarmx}.}

\item{...}{further arguments.}
}
\description{
Print Method for Object of Class \code{dtvarmx}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
