% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDTVARMx-methods-fit-dt-var-mx.R
\name{print.fitdtvarmx}
\alias{print.fitdtvarmx}
\title{Print Method for Object of Class \code{fitdtvarmx}}
\usage{
\method{print}{fitdtvarmx}(x, ...)
}
\arguments{
\item{x}{an object of class \code{fitdtvarmx}.}

\item{...}{further arguments.}
}
\description{
Print Method for Object of Class \code{fitdtvarmx}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
