% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDTVARMx-methods-fit-dt-var-mx.R
\name{summary.dtvarmx}
\alias{summary.dtvarmx}
\title{Summary Method for Object of Class \code{dtvarmx}}
\usage{
\method{summary}{dtvarmx}(object, ...)
}
\arguments{
\item{object}{an object of class \code{dtvarmx}.}

\item{...}{further arguments.}
}
\description{
Summary Method for Object of Class \code{dtvarmx}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
