% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDTVARMx-methods-fit-dt-var-id-mx.R
\name{summary.fitdtvaridmx}
\alias{summary.fitdtvaridmx}
\title{Summary Method for Object of Class \code{fitdtvaridmx}}
\usage{
\method{summary}{fitdtvaridmx}(object, means = TRUE, ...)
}
\arguments{
\item{object}{an object of class \code{fitdtvaridmx}.}

\item{means}{Logical.
If \code{means = TRUE}, return means.
Otherwise, the function returns raw estimates.}

\item{...}{further arguments.}
}
\description{
Summary Method for Object of Class \code{fitdtvaridmx}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
