% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDTVARMx-methods-fit-dt-var-mx.R
\name{summary.fitdtvarmx}
\alias{summary.fitdtvarmx}
\title{Summary Method for Object of Class \code{fitdtvarmx}}
\usage{
\method{summary}{fitdtvarmx}(object, ...)
}
\arguments{
\item{object}{an object of class \code{fitdtvarmx}.}

\item{...}{further arguments.}
}
\description{
Summary Method for Object of Class \code{fitdtvarmx}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
